Added the checksum back

+++ OutboundPacket.java
@@ -1,15 +1,15 @@
 import java.net.DatagramPacket;
 
 public class OutboundPacket {
-    private final int _sequenceNumber;
+    private final short _sequenceNumber;
     private final DatagramPacket _packet;
 
-    public OutboundPacket(int sequenceNumber, DatagramPacket packet) {
+    public OutboundPacket(short sequenceNumber, DatagramPacket packet) {
         _sequenceNumber = sequenceNumber;
         _packet = packet;
     }
 
-    public int getSequenceNumber() {
+    public short getSequenceNumber() {
         return _sequenceNumber;
     }
